feat: Adds failover_connections to mongodbatlas_stream_connection

+++ internal/service/streamconnection/resource_stream_connection.go
@@ -14,6 +14,7 @@ import (
 	"github.com/mongodb/terraform-provider-mongodbatlas/internal/common/conversion"
 	"github.com/mongodb/terraform-provider-mongodbatlas/internal/common/validate"
 	"github.com/mongodb/terraform-provider-mongodbatlas/internal/config"
+	"go.mongodb.org/atlas-sdk/v20250312021/admin"
 )
 
 const streamConnectionName = "stream_connection"
@@ -73,7 +74,53 @@ type TFStreamConnectionCommonModel struct {
 	SchemaRegistryProvider       types.String `tfsdk:"schema_registry_provider"`
 	SchemaRegistryURLs           types.List   `tfsdk:"schema_registry_urls"`
 	SchemaRegistryAuthentication types.Object `tfsdk:"schema_registry_authentication"`
+	FailoverConnections          types.List   `tfsdk:"failover_connections"`
 }
+
+// TFFailoverConnectionModel represents one failover connection configuration within a stream connection.
+type TFFailoverConnectionModel struct {
+	ID                           types.String `tfsdk:"id"`
+	Name                         types.String `tfsdk:"name"`
+	Type                         types.String `tfsdk:"type"`
+	ClusterName                  types.String `tfsdk:"cluster_name"`
+	ClusterProjectID             types.String `tfsdk:"cluster_project_id"`
+	DBRoleToExecute              types.Object `tfsdk:"db_role_to_execute"`
+	BootstrapServers             types.String `tfsdk:"bootstrap_servers"`
+	Authentication               types.Object `tfsdk:"authentication"`
+	Config                       types.Map    `tfsdk:"config"`
+	Security                     types.Object `tfsdk:"security"`
+	Networking                   types.Object `tfsdk:"networking"`
+	AWS                          types.Object `tfsdk:"aws"`
+	Azure                        types.Object `tfsdk:"azure"`
+	GCP                          types.Object `tfsdk:"gcp"`
+	URL                          types.String `tfsdk:"url"`
+	Headers                      types.Map    `tfsdk:"headers"`
+	SchemaRegistryProvider       types.String `tfsdk:"schema_registry_provider"`
+	SchemaRegistryURLs           types.List   `tfsdk:"schema_registry_urls"`
+	SchemaRegistryAuthentication types.Object `tfsdk:"schema_registry_authentication"`
+}
+
+var FailoverConnectionObjectType = types.ObjectType{AttrTypes: map[string]attr.Type{
+	"id":                            types.StringType,
+	"name":                          types.StringType,
+	"type":                          types.StringType,
+	"cluster_name":                  types.StringType,
+	"cluster_project_id":            types.StringType,
+	"db_role_to_execute":            DBRoleToExecuteObjectType,
+	"bootstrap_servers":             types.StringType,
+	"authentication":                ConnectionAuthenticationObjectType,
+	"config":                        types.MapType{ElemType: types.StringType},
+	"security":                      ConnectionSecurityObjectType,
+	"networking":                    NetworkingObjectType,
+	"aws":                           AWSObjectType,
+	"azure":                         AzureObjectType,
+	"gcp":                           GCPObjectType,
+	"url":                           types.StringType,
+	"headers":                       types.MapType{ElemType: types.StringType},
+	"schema_registry_provider":      types.StringType,
+	"schema_registry_urls":          types.ListType{ElemType: types.StringType},
+	"schema_registry_authentication": SchemaRegistryAuthenticationObjectType,
+}}
 
 type TFStreamConnectionModel struct {
 	TFStreamConnectionCommonModel
@@ -259,6 +306,45 @@ func (r *streamConnectionRS) Create(ctx context.Context, req resource.CreateRequ
 		resp.Diagnostics.Append(diags...)
 		return
 	}
+
+	// Write partial state so the primary connection isn't orphaned if failover creation fails.
+	newStreamConnectionModel.FailoverConnections = types.ListNull(FailoverConnectionObjectType)
+	resp.Diagnostics.Append(resp.State.Set(ctx, newStreamConnectionModel)...)
+	if resp.Diagnostics.HasError() {
+		return
+	}
+
+	// Create failover connections if configured.
+	if !streamConnectionPlan.FailoverConnections.IsNull() && !streamConnectionPlan.FailoverConnections.IsUnknown() && len(streamConnectionPlan.FailoverConnections.Elements()) > 0 {
+		var planFCs []TFFailoverConnectionModel
+		resp.Diagnostics.Append(streamConnectionPlan.FailoverConnections.ElementsAs(ctx, &planFCs, false)...)
+		if resp.Diagnostics.HasError() {
+			return
+		}
+		conns := make([]admin.StreamsConnection, len(planFCs))
+		for i, fc := range planFCs {
+			fcItem := fc
+			conn, diags := newFailoverConnectionReq(ctx, &fcItem)
+			if diags.HasError() {
+				resp.Diagnostics.Append(diags...)
+				return
+			}
+			conns[i] = *conn
+		}
+		_, _, err := connV2.StreamsApi.CreateFailoverConnections(ctx, projectID, workspaceOrInstanceName, connectionName, &admin.StreamsCreateFailoverConnectionsRequest{Connections: conns}).Execute()
+		if err != nil {
+			resp.Diagnostics.AddError("error creating failover connections", err.Error())
+			return
+		}
+	}
+
+	// Read back the failover connections (includes computed id values).
+	fcList, diags := newTFFailoverConnectionsList(ctx, connV2.StreamsApi, projectID, workspaceOrInstanceName, streamConnectionPlan.FailoverConnections, types.ListNull(FailoverConnectionObjectType))
+	if diags.HasError() {
+		resp.Diagnostics.Append(diags...)
+		return
+	}
+	newStreamConnectionModel.FailoverConnections = fcList
 	resp.Diagnostics.Append(resp.State.Set(ctx, newStreamConnectionModel)...)
 }
 
@@ -294,6 +380,14 @@ func (r *streamConnectionRS) Read(ctx context.Context, req resource.ReadRequest,
 		resp.Diagnostics.Append(diags...)
 		return
 	}
+
+	fcList, diags := newTFFailoverConnectionsList(ctx, connV2.StreamsApi, projectID, workspaceOrInstanceName, streamConnectionState.FailoverConnections, streamConnectionState.FailoverConnections)
+	if diags.HasError() {
+		resp.Diagnostics.Append(diags...)
+		return
+	}
+	newStreamConnectionModel.FailoverConnections = fcList
+
 	resp.Diagnostics.Append(resp.State.Set(ctx, newStreamConnectionModel)...)
 }
 
@@ -341,6 +435,25 @@ func (r *streamConnectionRS) Update(ctx context.Context, req resource.UpdateRequ
 		resp.Diagnostics.Append(diags...)
 		return
 	}
+
+	// Sync failover connections: create new ones, update changed ones, delete removed ones.
+	var streamConnectionState TFStreamConnectionModel
+	resp.Diagnostics.Append(req.State.Get(ctx, &streamConnectionState)...)
+	if resp.Diagnostics.HasError() {
+		return
+	}
+	if diags := syncFailoverConnections(ctx, connV2.StreamsApi, projectID, workspaceOrInstanceName, connectionName, streamConnectionPlan.FailoverConnections, streamConnectionState.FailoverConnections); diags.HasError() {
+		resp.Diagnostics.Append(diags...)
+		return
+	}
+
+	fcList, diags := newTFFailoverConnectionsList(ctx, connV2.StreamsApi, projectID, workspaceOrInstanceName, streamConnectionPlan.FailoverConnections, streamConnectionState.FailoverConnections)
+	if diags.HasError() {
+		resp.Diagnostics.Append(diags...)
+		return
+	}
+	newStreamConnectionModel.FailoverConnections = fcList
+
 	resp.Diagnostics.Append(resp.State.Set(ctx, newStreamConnectionModel)...)
 }
 
@@ -359,6 +472,12 @@ func (r *streamConnectionRS) Delete(ctx context.Context, req resource.DeleteRequ
 		return
 	}
 	connectionName := streamConnectionState.ConnectionName.ValueString()
+
+	// Delete failover connections first before removing the primary connection.
+	if diags := deleteAllFailoverConnections(ctx, connV2.StreamsApi, projectID, instanceName, connectionName, streamConnectionState.FailoverConnections); diags.HasError() {
+		resp.Diagnostics.Append(diags...)
+		return
+	}
 
 	deleteTimeout, diags := streamConnectionState.Timeouts.Delete(ctx, DefaultConnectionTimeout)
 	resp.Diagnostics.Append(diags...)
